Split chunker into 2 separate modules (#1)
* split chunker file

* add new command param for file type and update readme

* update readme

* updated main to quit when passing in unsupported file types

---------
