+++ info/README.adoc
@@ -1,12 +1,12 @@
-= Using `java-code-chunker`
+= Using `data-chunker`
 
 To use the Python package (as an example), use the following command:
 
-    $ python3 main.py training/test
+    $ python3 main.py training/test java
 
 Where:
 
-* `training/test` is the directory with Java code in it in this example. Any directory with any number of subdirectories may be used. Note that `*.java` files will be searched for in this example.
+* `training/test` is the directory with the code in it in this example. Any directory with any number of subdirectories may be used. Note that `*.java` files will be searched for in this example.
 
 The output of the command will print to console something like this:
